+++ src/renderer/index.js
@@ -4,6 +4,7 @@ const remote = require('electron').remote
 const options = new Options(remote.process.argv)
 
 require('./patches/console')
+require('./patches/debug')({ isTTY: options.isTTY })
 require('./keyboard/bindings')
 
 const Cucumber = require('@cucumber/cucumber')