+++ src/renderer/api/AppElements.js
@@ -1,18 +1,21 @@
 module.exports = class AppElements {
   constructor() {
-    this.fakeBrowsers = []
+    /**
+     * @private
+     */
+    this.fakeBrowserElements = []
   }
 
   /**
    * Creates a new DOM node that looks like a browser window and adds it to the DOM.
    *
-   * @param document the DOM document
-   * @param title the title of the top bar
-   * @return a DOM node that can be used to mount a DOM application such as a React or Vue element.
+   * @param {Document} document the DOM document
+   * @param {string} title the title of the top bar
+   * @return {Element} a DOM node that can be used to mount a DOM application such as a React or Vue element.
    */
   create(document, title) {
     const fakeBrowserElement = document.createElement('div')
-    this.fakeBrowsers.push(fakeBrowserElement)
+    this.fakeBrowserElements.push(fakeBrowserElement)
     fakeBrowserElement.innerHTML = `<div class="cucumber-electron-fake-browser">
       <div class="cucumber-electron-fake-browser-top">
         <span class="cucumber-electron-fake-browser-dot cucumber-electron-fake-browser-red"></span>
@@ -34,6 +37,6 @@ module.exports = class AppElements {
    * Destroys all previously created app elements.
    */
   destroyAll() {
-    this.fakeBrowsers.forEach(fakeBrowser => fakeBrowser.remove())
+    this.fakeBrowserElements.forEach(fakeBrowser => fakeBrowser.remove())
   }
 }