+++ @types/cucumber-electron/index.d.ts
@@ -1,0 +1,3 @@
+export var AppElements: {
+    new (): import("./AppElements");
+};