+++ src/renderer/api/AppElements.js
@@ -11,7 +11,7 @@ module.exports = class AppElements {
    *
    * @param {Document} document the DOM document
    * @param {string} title the title of the top bar
-   * @return {Element} a DOM node that can be used to mount a DOM application such as a React or Vue element.
+   * @return {HTMLElement} a DOM node that can be used to mount a DOM application such as a React or Vue element.
    */
   create(document, title) {
     const fakeBrowserElement = document.createElement('div')