+++ @types/cucumber-electron/fakeWindow.d.ts
@@ -1,0 +1,2 @@
+declare function _exports(document: Document, title: string, paneClass: string): HTMLElement;
+export = _exports;