+++ CHANGELOG.md
@@ -13,6 +13,11 @@ Please visit [cucumber/CONTRIBUTING.md](https://github.com/cucumber/cucumber/blo
 
 ### Added
 
+* Add an API that can be used in step definitions and hooks:
+  `const CucumberElectron = require('@cucumber/electron')`
+
+* Add `CucumberElectron#appendFakeBrowser(name: string)`
+
 ### Changed
 
 ### Deprecated