+++ CHANGELOG.md
@@ -13,7 +13,7 @@ Please visit [cucumber/CONTRIBUTING.md](https://github.com/cucumber/cucumber/blo
 
 ### Added
 
-* Add a FakeBrowsers API that can be used in step definitions and hooks
+* Add a AppElements API that can be used in step definitions and hooks
 
 ### Changed