+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Please visit [cucumber/CONTRIBUTING.md](https://github.com/cucumber/cucumber/blo
 
 ### Added
 
+* Restored support for the `debug` module.
 * Add a AppElements API that can be used in step definitions and hooks
 * Add TypeScript definitions for API